// frontend/packages/profile-ui/src/generated/models/ProfileListItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProfileListItem = {
    entity_id: string;
    entity_type_id: string;
    version: number;
    created_at: string;
    /**
     * Короткий preview текущего документа (для list UI).
     */
    preview: string;
};